Fix compilation when NDEBUG is defined
XORG_DEFAULT_OPTIONS enables -Werror=return-type

xcb_image_get_pixel() and xcb_image_create_native() assume assert(0) is
noreturn, and don't return a value in that case.  But assert(0) isn't noreturn
if NDEBUG is defined.

Fix this by returning a suitable value.

(Also, on the MinGW target, assert(0) is non-standard and may return even in the
!NDEBUG case, as the C runtime may ask the user if they want to abort or not...)

/jhbuild/checkout/xcb/xcb-util-image/image/xcb_image.c: In function 'xcb_image_get_pixel':
/jhbuild/checkout/xcb/xcb-util-image/image/xcb_image.c:754:1: error: control reaches end of non-void function [-Wreturn-type]
/jhbuild/checkout/xcb/xcb-util-image/image/xcb_image.c: In function 'xcb_image_create_native':
/jhbuild/checkout/xcb/xcb-util-image/image/xcb_image.c:190:1: error: control reaches end of non-void function [-Wreturn-type]

Signed-off-by: Jon TURNEY <jon.turney@dronecode.org.uk>
Signed-off-by: Peter Harris <pharris@opentext.com>